requirements.txt
torch==2.2.4+cu118
torchvision
numpy
Pillow
opencv-python
ffmpeg-python
tqdm
realesrgan
redis
